Profile: Game compiles, particles buggy
