+++ game/src/game/profiles/ProfileSystem.cpp
@@ -257,7 +257,7 @@ pip_t * ProfileSystem::pro_get_ppip( const PRO_REF iobj, int pip_index )
     return LOADED_PIP( local_pip ) ? PipStack.lst + local_pip : nullptr;
 }
 
-int ProfileSystem::loadOneProfile(const char* pathName, int slot_override )
+PRO_REF ProfileSystem::loadOneProfile(const char* pathName, int slot_override )
 {
     bool required = !(slot_override < 0 || slot_override >= INVALID_PRO_REF);
 
@@ -270,18 +270,18 @@ int ProfileSystem::loadOneProfile(const char* pathName, int slot_override )
         // The data file wasn't found
         if ( required )
         {
-            log_debug( "load_one_profile_vfs() - \"%s\" was not found. Overriding a global object?\n", pathName );
+            log_debug( "ProfileSystem::loadOneProfile() - \"%s\" was not found. Overriding a global object?\n", pathName );
         }
         else if ( required && slot_override > PMod->importamount * MAX_IMPORT_PER_PLAYER )
         {
-            log_warning( "load_one_profile_vfs() - Not able to open file \"%s\"\n", pathName );
+            log_warning( "ProfileSystem::loadOneProfile() - Not able to open file \"%s\"\n", pathName );
         }
 
         return INVALID_PRO_REF;
     }
 
     // convert the slot to a profile reference
-    PRO_REF iobj = ( PRO_REF )islot;
+    PRO_REF iobj = static_cast<PRO_REF>(islot);
 
     // throw an error code if we are trying to load over an existing profile
     // without permission
@@ -290,11 +290,11 @@ int ProfileSystem::loadOneProfile(const char* pathName, int slot_override )
         // Make sure global objects don't load over existing models
         if ( required && SPELLBOOK == iobj )
         {
-            log_error( "load_one_profile_vfs() - object slot %i is a special reserved slot number (cannot be used by %s).\n", SPELLBOOK, pathName );
+            log_error( "ProfileSystem::loadOneProfile() - object slot %i is a special reserved slot number (cannot be used by %s).\n", SPELLBOOK, pathName );
         }
         else if ( required && overrideslots )
         {
-            log_error( "load_one_profile_vfs() - object slot %i used twice (%s, %s)\n", REF_TO_INT( iobj ), _profilesLoaded[iobj]->getName().c_str(), pathName );
+            log_error( "ProfileSystem::loadOneProfile() - object slot %i used twice (%s, %s)\n", REF_TO_INT( iobj ), _profilesLoaded[iobj]->getFilePath().c_str(), pathName );
         }
         else
         {
@@ -303,8 +303,15 @@ int ProfileSystem::loadOneProfile(const char* pathName, int slot_override )
         }
     }
 
-    //Actually allocate the object and create it
-    _profilesLoaded[iobj] = std::make_shared<ObjectProfile>(pathName, islot);
+    std::shared_ptr<ObjectProfile> profile = ObjectProfile::loadFromFile(pathName, iobj);
+    if(!profile)
+    {
+        log_warning("ProfileSystem::loadOneProfile() - Failed to load (%s) into slot number %d\n", pathName, iobj);
+        return INVALID_PRO_REF;
+    }
+
+    //Success! Store object into the loaded profile map
+    _profilesLoaded[iobj] = profile;
 
     //ZF> TODO: This is kind of a dirty hack and could be done cleaner. If this item is the book object, 
     //    then the icons are also loaded into the global book icon array
@@ -323,4 +330,10 @@ TX_REF ProfileSystem::getSpellBookIcon(size_t index) const
 {
     if(_bookIcons.empty()) return INVALID_TX_REF;
     return _bookIcons[index % _bookIcons.size()];
+}
+
+bool ProfileSystem::exportProfileToFile(const std::string &filePath, const CHR_REF character)
+{
+    //TODO: not implemented
+    log_error("ProfileSystem::exportProfileToFile() - unimplemented function, not ported yet\n");
 }
